♻️ Create a SessionStorage interface
