+++ packages/core/src/domain/session/sessionStorage.ts
@@ -1,0 +1,14 @@
+export interface SessionState {
+  id?: string
+  created?: string
+  expire?: string
+  lock?: string
+
+  [key: string]: string | undefined
+}
+
+export interface SessionStorage {
+  persistSession: (session: SessionState) => void
+  retrieveSession: () => SessionState
+  clearSession: () => void
+}
